STR Sink or Float response
Added a detailed explanation as well as a link to a research paper that tackled a somewhat similar problem to the "Does it sink or float?" paradox for Special relativity in the intuition category. The previous commit was reverted owing to some calculation errors that creeped in for God-knows what reason!